Убран ненужный файл
